==> 8. Use Cases/customer-churn-data-preparation/01_read_dataset.ipynb <==
# 1. The impact of synthetic data on customer churn

We are all familiar with the axiom "Garbage in, garbage out", and this is very much true, specially in a setting and market where we see models getting more and more commmoditized. The business advatage will remain in the component that it is unique to every organization - the data.

In every case - and particularly for credit scoring use cases - data preparation is a paramount. Nevertheless, and althoug the achievments we have observed in the past few years, data preparation is still the most challenging and time-consuming step. Ensuring data quality helps data teams to achieve bigger ROI from AI initiatives at a fraction of the effort it used to, translating into better scorecards that positively impact the business and customer experience.

When we look into the credit scoring, there are particular issues that can dampen model accuracy - presence of outliers, missing values and the presence of imbalanced classes.

In this usecase we will explore not only an iterative, traceable and comparable data processing for to improve the quality of the data for credit scorecards, but also how to mitigate each one of the identified challenges: missing data, presence of duplicates and last but not the least, imabalanced data.

The dataset leveraged for the use-case can be easily found in [Kaggle - Telco churn](https://www.kaggle.com/datasets/blastchar/telco-customer-churn/code).


### Import needed packages

import os
import pandas as pd
from pathlib import Path

# Importing YData's SDK packages
from ydata.labs.datasources import DataSources
from ydata.dataset import Dataset

from functions.saving_functions import save_file

## Read the data

The first step is to read the data. We have previously created the DataSource at the level of the UI. 

Because Fabric enables an ease integration between the labs and other workbenck elements (such as Datasources and Synthetisizers). The code below depicts how:

# Importing YData's packages
from ydata.labs import DataSources
# Reading the Dataset from the DataSource
datasource = DataSources.get(uid='17488144-9467-4902-8df9-5380a6a5700e', namespace='a0717134-f616-42fd-a38e-468c63d26802')
dataset = datasource.dataset
# Getting the calculated Metadata to get the profile overview information in the labs
metadata = datasource.metadata
print(metadata)

#Convert dataset to Dask engine
dd_dataset = dataset.to_dask()

#Rename the columns tenure and Churn Value for ease of exploration
dd_dataset = dd_dataset.rename(columns={"Tenure Months": "Tenure", "Churn Value": "Churn"})

dataset = Dataset(dd_dataset)

## Data Exploration 

As a critical first step, we should develop a more comprehensive view of our data in order to understand the key drivers of customer churn.

# Quickly previewing the Dataset
dataset.head()

It is crucial for any Data Science process that the dataset variable types are correctly set, otherwise the process of data preparation might not be optimal, resulting in lower performance at the level of the classifiers to be built!

print(dataset)

Based on the above dataset summary and metadata information, we have identified that the **CustomerID** column is in fact an *ID* type. For that reason we have to update the Metadata selected data types. 

#Updating the metadata with the correct data types
metadata.update_datatypes({'CustomerID': 'string',
                           'Churn Reason': 'longtext'})

## Create pipeline outputs

metadata.save('metadata.pkl')
#saving the dataset
save_file(dataset, file_path='dataset.pkl')

#Get metadata warnings
warnings=[]
for warning, val in metadata.warnings.items():
    for col in val:
        try:
            level = col.details['level'].name
            value = round(col.details['value'], 4)
        except:
            level = None
            value = None
        warnings.append({'warning': warning, 'column': col.column, 'level': level, 'value':value})

df_warnings= pd.DataFrame(warnings)

import json

metadata = {
    'outputs' : [
        {
      'type': 'table',
      'storage': 'inline',
      'format': 'csv',
      'header': list(df_warnings.columns),
      'source': df_warnings.to_csv(header=False, index=False)
    }
    ]
  }

with open('mlpipeline-ui-metadata.json', 'w') as metadata_file:
    json.dump(metadata, metadata_file)
