==> 8. Use Cases/customer-churn-data-preparation/03_data_preparation.ipynb <==
# 3. Data preprocessing

Feature engineering can be really helpful for data science projects by allowing the extraction/selection of the most relevant and important variables for a use case. Optimizing feature selection based on your downstream application is important to expective the expected results of AI developed solutions.

In what concerns this customer churn use case, there is a lot of potential around location based columns and the available target information.

### Import needed packages

import os
from pathlib import Path

from functions.saving_functions import save_file, read_file

from pandas_profiling import ProfileReport

from ydata.dataset import Dataset
from ydata.metadata import Metadata

from sklearn.cluster import DBSCAN
import joblib

#### Environment variables

feat_eng = bool(os.environ.get('FEATURE_ENG', "True"))

## Read the dataset & load the metadata

dataset = read_file('dataset.pkl')
metadata = Metadata.load('metadata.pkl')

dataset.head()

print(metadata)

## Preprocess & Feature Engineering

#Remove any ID columns as they are expected to be unique throughout the whoel dataset
dataset = dataset.drop_columns(metadata.id_vars)

#Clean the dataset based on Metadata
dataset_cols = list(dataset.columns)
if len(metadata.warnings['constant']) > 0:
    const_cols = [warning.column for warning in metadata.warnings['constant']]

dataset = dataset.select_columns(list(set(dataset_cols)-set(const_cols)))

len_dataset = len(dataset)
cardinality_cols = [(warning.details['value']/len_dataset) for warning in metadata.warnings['cardinality']]

#Lat lon is a calculated field which means it can be removed

#Let's have a look into the cities representative distribution
cities_cardinality = dataset.to_dask().groupby('City').count().sort_values(by='Zip Code', ascending=False)[['Zip Code']].compute()

## Explore the city impact with compare profiling

ten_cust = list(cities_cardinality[cities_cardinality['Zip Code'] <=10].reset_index()['City'])
plus10_cust = list(cities_cardinality[cities_cardinality['Zip Code'] >10].reset_index()['City'])

small_cities = dataset._data[dataset._data['City'].isin(ten_cust)].compute()
bigger_cities = dataset._data[dataset._data['City'].isin(plus10_cust)].compute()

biggercities_r = ProfileReport(bigger_cities, title='Cities with +customers')
smallcities_r = ProfileReport(small_cities, title='Cities with less customers')

compare_report = biggercities_r.compare(smallcities_r)

#### Dropping columns that are not relevant for the use case

## Drop variables related to location as they seem to hold no information
location_var = ['Latitude', 'Longitude', 'Lat Long']

flrt_dataset = dataset.select_columns(list(set(dataset.columns) - set(location_var)))

#create a new column based on wether is Los Angeles or not
flrt_dataset.to_dask()['City'].unique().compute()

flrt_dataset.head()

dd_data = flrt_dataset.to_dask()

dd_city = dd_data.groupby(['City'])[['Monthly Charges','Total Charges']].mean().reset_index()
dd_city.columns = ['City', 'Avg Montly charges per city','Avg Total charges per city']

#Add new columns to extract the city information
dd_data = dd_data.merge(dd_city,on='City',suffixes=('_x', '_y'))

dd_data = dd_data.drop(['City', 'Churn Label'], axis=1)

### Dropping churn reason as it leaks information

dd_data = dd_data.drop('Churn Reason', axis=1)

## Missing values processing

dd_data = dd_data.dropna()

preprocessed_data = Dataset(dd_data)
preprocessed_meta = Metadata(preprocessed_data)

## Save as new datasource

## add here the connector details
from ydata.labs.connectors import Connectors
from ydata.labs import GoogleCloudStorageDataSource, DataType, FileType

conn=Connectors.get('{connector-id}')

conn.write_file(data=preprocessed_data, 
                path="{add-file-path}")

datasource = GoogleCloudStorageDataSource(name='Customer churn - Preprocessed',
                                          connector=conn,
                                          data_type=DataType.TABULAR,
                                          path="{add-file-path}",
                                          file_type=FileType.CSV,
                                         )
datasource.create()

## Create pipeline outputs

### Saving the artifacts

parameters = {'preprocessed_id': datasource.id}

metadata = preprocessed_meta.save('proc_metadata.pkl')

#Saving pipeline parameters file
save_file(parameters, file_path='pipeline_parameters.pkl')

import json

profile_pipeline_output = {
    'outputs' : [
        {
      'type': 'web-app',
      'storage': 'inline',
      'source': compare_report.to_html(),
    }
    ]
  }

with open('mlpipeline-ui-metadata.json', 'w') as metadata_file:
    json.dump(profile_pipeline_output, metadata_file)
