==> 8. Use Cases/customer-churn-data-preparation/04_profiling_compare.ipynb <==
# 2. Profiling - Compare data after preprocessing

The first step in a data preparation pipeline is the exploratory data analysis (EDA). In a nutshell, data exploration and data cleansing are hand-to-hand and both are mutually iterative steps.

*But what does data exploration includes? And how to make a better data exploration giving we are building a customer churn model classifier?*

Data exploration includes both univariate and bivariate analysis and ranges from univariate statistics and frequency distributions to correlations, cross-tabulation, and characteristic analysis. add here detail about pandas-profiling and data exploration in general (re-use the sentence above)


### Import needed packages

import os
import json
import warnings
import pandas as pd
from pathlib import Path

from ydata.labs import DataSources

from ydata.profiling import ProfileReport

from functions.saving_functions import read_file

pipeline_parameters = read_file(file_path='pipeline_parameters.pkl')

og_source = os.environ.get('OG_Datasource_id', "{add-datasource-id}")
proc_source = pipeline_parameters['preprocessed_id']

title_og = os.environ.get('ORIGINAL_TITLE', "Original dataset")
title_comp = os.environ.get('COMPARE_TITLE', "Processed dataset")

## Read the sources of data

#get the dataset from the reading step
og_dataset = DataSources.get(uid=og_source).read().to_pandas()
proc_dataset = DataSources.get(uid=proc_source).read().to_pandas()

## Profiling the data

profile_title = title_og
compare_title = title_comp

og_profile = ProfileReport(df=og_dataset, title=profile_title)

compare_profile = ProfileReport(df=proc_dataset, title=compare_title)

compare = og_profile.compare(compare_profile)

## Create pipeline outputs

import json

profile_pipeline_output = {
    'outputs' : [
        {
      'type': 'web-app',
      'storage': 'inline',
      'source': compare.to_html(),
    }
    ]
  }

with open('mlpipeline-ui-metadata.json', 'w') as metadata_file:
    json.dump(profile_pipeline_output, metadata_file)
