==> 8. Use Cases/customer-churn-data-preparation/05_holdout.ipynb <==
# 4. Holdout creation

### Import the needed packages

import os
from ydata.labs.datasources import DataSources

from ydata.dataset import Dataset
from ydata.metadata import Metadata
from ydata.dataset.holdout import Holdout

from functions.saving_functions import read_file, save_file

#### Get environment variables values

pipeline_parameters = read_file(file_path='pipeline_parameters.pkl')

### Read the dataset & Metadata

metadata = Metadata.load('proc_metadata.pkl')

dataset = DataSources.get(uid=pipeline_parameters['preprocessed_id']).read()

dataset.head()

## Get an holdout sample

#Creating the holdout set
#Reading the holdout size from the pipeline definition
holdout_ratio = float(os.environ.get('HOLDOUT_RATIO', 0.15))

data = dataset.to_pandas()
holdout_set = data.sample(frac=holdout_ratio)
train_set = data.drop(holdout_set.index)

train_set = Dataset(train_set)
holdout_set = Dataset(holdout_set)

#recalculate the metadata
metadata_train = Metadata(train_set)
metadata_holdout = Metadata(holdout_set)

metadata_holdout = Metadata(holdout_set)

metadata_holdout

print(metadata_holdout)

## Create pipeline outputs

pipeline_parameters= read_file('pipeline_parameters.pkl')
pipeline_parameters['holdout'] = holdout_ratio
save_file(pipeline_parameters, 'pipeline_parameters.pkl')

metadata_train.save('metadata_train.pkl')
metadata_holdout.save('metadata_holdout.pkl')

save_file(train_set, 'train_set.pkl')
save_file(holdout_set, 'holdout_set.pkl')
