==> 8. Use Cases/customer-churn-data-preparation/06_data_augmentation.ipynb <==
# 6. Augment underepresented class with Synthetic data

### Import the need packages

import os
from pathlib import Path
import pandas as pd
import numpy as np

import dask.dataframe as dd

from ydata.profiling import ProfileReport

from ydata.synthesizers.regular import RegularSynthesizer

from functions.saving_functions import read_file, save_file

### Getting the running settings

#Reading the environment variables
augment = int(os.environ.get('AUGMENT', 1))
augment_class = os.environ.get('AUGMENT_CLASS', 'Churn')
augment_class_val = int(os.environ.get('AUGMENT_CLASS_VAL', 1))
sample_size = int(os.environ.get('SAMPLE_SIZE', 1000))

try:
    augment_class_val = int(os.environ.get('AUGMENT_CLASS_VAL', 1))
except: 
    augment_class_val = os.environ.get('AUGMENT_CLASS_VAL', 1)

## Read the dataset & Metadata

metadata = read_file('metadata_train.pkl')
dataset = read_file('train_set.pkl')

# Importing YData's packages
from ydata.labs import DataSources
# Reading the Dataset from the DataSource
datasource = DataSources.get(uid='{datasource-id}')

dataset = datasource.dataset
# Getting the calculated Metadata to get the profile overview information in the labs
metadata = datasource.metadata

### The data synthesis

#### Training the synthesizer

from ydata.dataset import Dataset
from ydata.metadata import Metadata 

if bool(augment):    
    synth = RegularSynthesizer()
    synth.fit(dataset, metadata=metadata, condition_on=[augment_class])
    synth_sample = synth.sample(sample_size)
    
    #Convert synth sample to DASK
    synth_dataset = synth_sample.to_dask()
    
    #Re-compute dataset & Metadata    
    dataset = Dataset(dd.concat([dataset.to_dask(), synth_dataset]))
    metadata = Metadata(dataset)

## Pipeline Outputs

#Saving the generated sample
if augment:
    save_file(dataset, 'train_set.pkl')
    metadata.save('metadata_train.pkl')

pipeline_parameters = read_file('pipeline_parameters.pkl')
pipeline_parameters['augment'] = augment
pipeline_parameters['sample_size'] = sample_size
save_file(pipeline_parameters, 'pipeline_parameters.pkl')
